requirements.txt
streamlit
google-generativeai
python-docx 
pymupdf 
pdfplumber 
python-dotenv

